feat: use AWS-specific content type within the response of all JSON-based controllers
