+++ src/main/java/io/github/hectorvent/floci/core/common/AwsJson11Controller.java
@@ -1,5 +1,7 @@
 package io.github.hectorvent.floci.core.common;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.hectorvent.floci.services.acm.AcmJsonHandler;
 import io.github.hectorvent.floci.services.ecs.EcsJsonHandler;
 import io.github.hectorvent.floci.services.apigatewayv2.ApiGatewayV2JsonHandler;
@@ -10,13 +12,10 @@ import io.github.hectorvent.floci.services.kinesis.KinesisJsonHandler;
 import io.github.hectorvent.floci.services.kms.KmsJsonHandler;
 import io.github.hectorvent.floci.services.secretsmanager.SecretsManagerJsonHandler;
 import io.github.hectorvent.floci.services.ssm.SsmJsonHandler;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.HttpHeaders;
-import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import org.jboss.logging.Logger;
 
@@ -83,7 +82,7 @@ public class AwsJson11Controller {
 
     @POST
     @Consumes("application/x-amz-json-1.1")
-    @Produces(MediaType.APPLICATION_JSON)
+    @Produces("application/x-amz-json-1.1")
     public Response handle(
             @HeaderParam("X-Amz-Target") String target,
             @Context HttpHeaders httpHeaders,
@@ -127,10 +126,7 @@ public class AwsJson11Controller {
             prefix = ECS_TARGET_PREFIX;
             serviceName = "ECS";
         } else {
-            return Response.status(400)
-                    .entity(new AwsErrorResponse("UnknownOperationException",
-                            "Unknown operation: " + target))
-                    .build();
+            return JsonErrorResponseUtils.createUnknownOperationErrorResponse(target);
         }
 
         String action = target.substring(prefix.length());
@@ -154,16 +150,11 @@ public class AwsJson11Controller {
                 default -> null;
             };
         } catch (AwsException e) {
-            return Response.status(e.getHttpStatus())
-                    .type(MediaType.APPLICATION_JSON)
-                    .entity(new AwsErrorResponse(e.getErrorCode(), e.getMessage()))
-                    .build();
+            return JsonErrorResponseUtils.createErrorResponse(e);
         } catch (Exception e) {
             LOG.errorf("Error processing %s request", serviceName, e);
-            return Response.status(500)
-                    .type(MediaType.APPLICATION_JSON)
-                    .entity(new AwsErrorResponse("InternalServerError", e.getMessage()))
-                    .build();
+            return JsonErrorResponseUtils.createErrorResponse(e);
         }
     }
+
 }
