+++ src/main/java/io/github/hectorvent/floci/services/apigatewayv2/ApiGatewayV2JsonHandler.java
@@ -1,15 +1,14 @@
 package io.github.hectorvent.floci.services.apigatewayv2;
 
-import io.github.hectorvent.floci.core.common.AwsErrorResponse;
 import io.github.hectorvent.floci.core.common.AwsException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import io.github.hectorvent.floci.core.common.JsonErrorResponseUtils;
 import io.github.hectorvent.floci.services.apigatewayv2.model.*;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 import java.util.List;
@@ -51,15 +50,10 @@ public class ApiGatewayV2JsonHandler {
                 case "DeleteStage" -> handleDeleteStage(request, region);
                 case "CreateDeployment" -> handleCreateDeployment(request, region);
                 case "GetDeployments" -> handleGetDeployments(request, region);
-                default -> Response.status(400)
-                        .entity(new AwsErrorResponse("UnsupportedOperation",
-                                "Operation " + action + " is not supported."))
-                        .type(MediaType.APPLICATION_JSON).build();
+                default -> JsonErrorResponseUtils.createUnknownOperationErrorResponse(action);
             };
         } catch (AwsException e) {
-            return Response.status(e.getHttpStatus())
-                    .entity(new AwsErrorResponse(e.getErrorCode(), e.getMessage()))
-                    .type(MediaType.APPLICATION_JSON).build();
+            return JsonErrorResponseUtils.createErrorResponse(e);
         }
     }
 
@@ -69,7 +63,7 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Api api = service.createApi(region, map);
-        return Response.status(201).entity(toApiNode(api).toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toApiNode(api).toString()).build();
     }
 
     private Response handleGetApis(String region) {
@@ -77,13 +71,12 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         apis.forEach(a -> items.add(toApiNode(a)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     private Response handleGetApi(JsonNode request, String region) {
         String apiId = request.path("ApiId").asText();
-        return Response.ok(toApiNode(service.getApi(region, apiId)).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(toApiNode(service.getApi(region, apiId)).toString()).build();
     }
 
     private Response handleDeleteApi(JsonNode request, String region) {
@@ -99,15 +92,13 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Authorizer auth = service.createAuthorizer(region, apiId, map);
-        return Response.status(201).entity(toAuthorizerNode(auth).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toAuthorizerNode(auth).toString()).build();
     }
 
     private Response handleGetAuthorizer(JsonNode request, String region) {
         String apiId = request.path("ApiId").asText();
         String authorizerId = request.path("AuthorizerId").asText();
-        return Response.ok(toAuthorizerNode(service.getAuthorizer(region, apiId, authorizerId)).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(toAuthorizerNode(service.getAuthorizer(region, apiId, authorizerId)).toString()).build();
     }
 
     private Response handleGetAuthorizers(JsonNode request, String region) {
@@ -116,7 +107,7 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         authorizers.forEach(a -> items.add(toAuthorizerNode(a)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     private Response handleDeleteAuthorizer(JsonNode request, String region) {
@@ -133,15 +124,13 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Route route = service.createRoute(region, apiId, map);
-        return Response.status(201).entity(toRouteNode(route).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toRouteNode(route).toString()).build();
     }
 
     private Response handleGetRoute(JsonNode request, String region) {
         String apiId = request.path("ApiId").asText();
         String routeId = request.path("RouteId").asText();
-        return Response.ok(toRouteNode(service.getRoute(region, apiId, routeId)).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(toRouteNode(service.getRoute(region, apiId, routeId)).toString()).build();
     }
 
     private Response handleGetRoutes(JsonNode request, String region) {
@@ -150,7 +139,7 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         routes.forEach(r -> items.add(toRouteNode(r)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     private Response handleDeleteRoute(JsonNode request, String region) {
@@ -167,15 +156,13 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Integration integration = service.createIntegration(region, apiId, map);
-        return Response.status(201).entity(toIntegrationNode(integration).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toIntegrationNode(integration).toString()).build();
     }
 
     private Response handleGetIntegration(JsonNode request, String region) {
         String apiId = request.path("ApiId").asText();
         String integrationId = request.path("IntegrationId").asText();
-        return Response.ok(toIntegrationNode(service.getIntegration(region, apiId, integrationId)).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(toIntegrationNode(service.getIntegration(region, apiId, integrationId)).toString()).build();
     }
 
     private Response handleGetIntegrations(JsonNode request, String region) {
@@ -184,7 +171,7 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         integrations.forEach(i -> items.add(toIntegrationNode(i)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     // ──────────────────────────── Stage ────────────────────────────
@@ -194,15 +181,13 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Stage stage = service.createStage(region, apiId, map);
-        return Response.status(201).entity(toStageNode(stage).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toStageNode(stage).toString()).build();
     }
 
     private Response handleGetStage(JsonNode request, String region) {
         String apiId = request.path("ApiId").asText();
         String stageName = request.path("StageName").asText();
-        return Response.ok(toStageNode(service.getStage(region, apiId, stageName)).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(toStageNode(service.getStage(region, apiId, stageName)).toString()).build();
     }
 
     private Response handleGetStages(JsonNode request, String region) {
@@ -211,7 +196,7 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         stages.forEach(s -> items.add(toStageNode(s)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     private Response handleDeleteStage(JsonNode request, String region) {
@@ -228,8 +213,7 @@ public class ApiGatewayV2JsonHandler {
         @SuppressWarnings("unchecked")
         Map<String, Object> map = objectMapper.convertValue(request, Map.class);
         Deployment deployment = service.createDeployment(region, apiId, map);
-        return Response.status(201).entity(toDeploymentNode(deployment).toString())
-                .type(MediaType.APPLICATION_JSON).build();
+        return Response.status(201).entity(toDeploymentNode(deployment).toString()).build();
     }
 
     private Response handleGetDeployments(JsonNode request, String region) {
@@ -238,7 +222,7 @@ public class ApiGatewayV2JsonHandler {
         ObjectNode root = objectMapper.createObjectNode();
         ArrayNode items = root.putArray("Items");
         deployments.forEach(d -> items.add(toDeploymentNode(d)));
-        return Response.ok(root.toString()).type(MediaType.APPLICATION_JSON).build();
+        return Response.ok(root.toString()).build();
     }
 
     // ──────────────────────────── Serializers ────────────────────────────
