+++ src/test/java/io/github/hectorvent/floci/services/acm/AcmImportExportTest.java
@@ -1,10 +1,8 @@
 package io.github.hectorvent.floci.services.acm;
 
 import io.github.hectorvent.floci.services.acm.model.KeyAlgorithm;
+import io.github.hectorvent.floci.testing.RestAssuredJsonUtils;
 import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
-import io.restassured.config.EncoderConfig;
-import io.restassured.http.ContentType;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.MethodOrderer;
@@ -38,9 +36,7 @@ class AcmImportExportTest {
 
     @BeforeAll
     void setupTestCertificates() {
-        RestAssured.config = RestAssured.config().encoderConfig(
-                EncoderConfig.encoderConfig()
-                        .encodeContentTypeAs(ACM_CONTENT_TYPE, ContentType.TEXT));
+        RestAssuredJsonUtils.configureAwsContentTypes();
 
         // Generate a valid test certificate using CertificateGenerator
         CertificateGenerator.GeneratedCertificate generated = certificateGenerator.generateCertificate(
