+++ src/test/java/io/github/hectorvent/floci/services/sqs/SqsJsonProtocolTest.java
@@ -1,9 +1,7 @@
 package io.github.hectorvent.floci.services.sqs;
 
+import io.github.hectorvent.floci.testing.RestAssuredJsonUtils;
 import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
-import io.restassured.config.EncoderConfig;
-import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -34,10 +32,7 @@ class SqsJsonProtocolTest {
 
     @BeforeAll
     static void configureRestAssured() {
-        RestAssured.config = RestAssured.config().encoderConfig(
-            EncoderConfig.encoderConfig()
-                .encodeContentTypeAs(CONTENT_TYPE, ContentType.TEXT)
-        );
+        RestAssuredJsonUtils.configureAwsContentTypes();
     }
 
     // --- Root-path JSON 1.0 (POST /) ---
